[Refactor] : add snap card view in eb uikit
Git : #36
